Remove Test files
